Try to fix issue importing schemas in explorer
