+++ src/api/getAppDataSchema.ts
@@ -1,3 +1,4 @@
+import { importSchema } from '../importSchema'
 import { MetaDataError } from '../consts'
 import { AnyAppDataDocVersion } from 'generatedTypes'
 
@@ -14,16 +15,5 @@ export async function getAppDataSchema(version: string): Promise<AnyAppDataDocVe
     // Wrapping @cowprotocol/app-data Error into CowError
     const error = e as Error
     throw new MetaDataError(error.message)
-  }
-}
-
-export async function importSchema(version: string): Promise<AnyAppDataDocVersion> {
-  if (!/^\d+\.\d+\.\d+$/.test(version)) {
-    throw new Error(`AppData version ${version} is not a valid version`)
-  }
-  try {
-    return await import(`../../schemas/v${version}.json`)
-  } catch (e) {
-    throw new Error(`AppData version ${version} doesn't exist`)
   }
 }
